# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lerank platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile (student onboarding data)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /consultants:
    get:
      operationId: listConsultants
      tags: [consultants]
      summary: List consultant companies with filters
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: degreeLevel
          in: query
          schema:
            type: string
            enum: [bachelor, master, phd]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [bestMatch, priceLow, rating, mostStudents]
      responses:
        "200":
          description: List of consultants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultantCompany"
  /consultants/{id}:
    get:
      operationId: getConsultant
      tags: [consultants]
      summary: Get consultant company details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultantCompany"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications for current student
      responses:
        "200":
          description: Student applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new application (hire consultant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application with milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
...
  /admin/companies:
    get:
      operationId: adminListCompanies
      tags: [admin]
      summary: Admin - list all companies
      responses:
        "200":
          description: All companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultantCompany"
    post:
      operationId: adminCreateCompany
      tags: [admin]
      summary: Admin - create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultantCompany"
  /admin/companies/{id}:
    put:
      operationId: adminUpdateCompany
      tags: [admin]
      summary: Admin - update company status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyRequest"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultantCompany"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [student, company_admin, super_admin]
        avatarUrl:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        profile:
          $ref: "#/components/schemas/StudentProfile"
          nullable: true
        companyId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, fullName, role, onboardingCompleted, createdAt]
    StudentProfile:
      type: object
      properties:
        ieltsScore:
          type: number
          nullable: true
        satScore:
          type: integer
          nullable: true
        toeflScore:
          type: integer
          nullable: true
        gpa:
          type: number
          nullable: true
        gpaScale:
          type: number
          nullable: true
        budgetMin:
          type: integer
          nullable: true
        budgetMax:
          type: integer
          nullable: true
        educationBudget:
          type: string
          nullable: true
        preferredCountries:
          type: array
          items:
            type: string
        preferredMajor:
          type: string
          nullable: true
        degreeLevel:
          type: string
          nullable: true
        additionalNotes:
          type: string
          nullable: true
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fullName:
          type: string
      required: [email, password, fullName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    UpdateProfileRequest:
      type: object
      properties:
        ieltsScore:
          type: number
          nullable: true
        satScore:
          type: integer
          nullable: true
        toeflScore:
          type: integer
          nullable: true
        gpa:
          type: number
          nullable: true
        gpaScale:
          type: number
          nullable: true
        budgetMin:
          type: integer
          nullable: true
        budgetMax:
          type: integer
          nullable: true
        educationBudget:
          type: string
          nullable: true
        preferredCountries:
          type: array
          items:
            type: string
        preferredMajor:
          type: string
          nullable: true
        degreeLevel:
          type: string
          nullable: true
        additionalNotes:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
    ConsultantCompany:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        rating:
          type: number
        studentsHelped:
          type: integer
        successRate:
          type: number
        priceMin:
          type: integer
        priceMax:
          type: integer
        specializedCountries:
          type: array
          items:
            type: string
        degreeLevels:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, rating, studentsHelped, successRate, priceMin, priceMax, specializedCountries, degreeLevels, isActive, createdAt]
    CreateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        priceMin:
          type: integer
        priceMax:
          type: integer
        specializedCountries:
          type: array
          items:
            type: string
        degreeLevels:
          type: array
          items:
            type: string
      required: [name, priceMin, priceMax, specializedCountries, degreeLevels]
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        isActive:
          type: boolean
        priceMin:
          type: integer
        priceMax:
          type: integer
    Application:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        serviceName:
          type: string
        status:
          type: string
          enum: [active, completed, refunded, disputed]
        progressPercent:
          type: number
        totalAmount:
          type: integer
        escrowStatus:
          type: string
          enum: [pending, held, released, refunded]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, studentId, companyId, companyName, serviceName, status, progressPercent, totalAmount, escrowStatus, createdAt, updatedAt]
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          properties:
            milestones:
              type: array
              items:
                $ref: "#/components/schemas/Milestone"
          required: [milestones]
    CreateApplicationRequest:
      type: object
      properties:
        companyId:
          type: integer
        serviceName:
          type: string
        totalAmount:
          type: integer
        templateId:
          type: integer
          nullable: true
      required: [companyId, serviceName, totalAmount]
    Milestone:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [not_started, in_progress, under_review, completed]
        orderIndex:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required: [id, applicationId, name, status, orderIndex, updatedAt]
...