# lib/api-zod/src/generated/types/applicationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lerank platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from "./application";
import type { Milestone } from "./milestone";

export type ApplicationDetail = Application & {
  milestones: Milestone[];
};