# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lerank platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  ieltsScore?: number | null;
  satScore?: number | null;
  toeflScore?: number | null;
  gpa?: number | null;
  gpaScale?: number | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  educationBudget?: string | null;
  preferredCountries?: string[];
  preferredMajor?: string | null;
  degreeLevel?: string | null;
  additionalNotes?: string | null;
  onboardingCompleted?: boolean;
}