# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lerank platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfile } from "./studentProfile";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  fullName: string;
  role: UserRole;
  avatarUrl?: string | null;
  onboardingCompleted: boolean;
  profile?: StudentProfile | null;
  companyId?: number | null;
  createdAt: Date;
}